small refactor of commands
