+++ src/workflow/workflow.rs
@@ -1,34 +1,9 @@
-use crate::args::ArgSupplier;
-use crate::args::UserConfigProvider;
-use crate::args::DefaultUserConfigProvider;
+use crate::args::{ArgSupplier, DefaultUserConfigProvider, ZatCommand};
 use crate::args::cli_arg_supplier::CliArgSupplier;
-use crate::args::{ZatCommand, ProcessTemplatesArgs, BootstrapProjectArgs};
-
-use crate::config::TargetDir;
+use crate::command::{BootstrapProject, ProcessTemplates};
 use crate::error::ZatAction;
 
-use crate::error::ZatError;
-use crate::templates::TemplateVariableProvider;
-use crate::templates::DefaultTemplateVariableProvider;
-use crate::templates::TemplateConfigValidator;
-use crate::templates::DefaultTemplateConfigValidator;
-use crate::templates::TemplateVariableReview;
 
-use crate::token_expander::ExpandFilters;
-use crate::token_expander::DefaultExpandFilters;
-
-use crate::processor::ProcessTemplates;
-use crate::processor::DefaultProcessTemplates;
-
-use crate::post_processor::PostProcessingHook;
-use crate::post_processor::ShellHook;
-
-use crate::logging::VerboseLogger;
-use crate::logging::Logger;
-use std::format as s;
-use std::todo;
-
-//TODO: Make this into a trait and impl so we can test it.
 pub struct Workflow;
 
 impl Workflow {
@@ -41,63 +16,13 @@ impl Workflow {
 
     match config.command {
       ZatCommand::Process(process_templates_args) => {
-        Workflow::process_templates(config_provider, process_templates_args)
+        ProcessTemplates::process(config_provider, process_templates_args)
       },
 
-      ZatCommand::Bootstrap(bootstrap_project_args) => Workflow::process_boostrap(bootstrap_project_args),
+      ZatCommand::Bootstrap(bootstrap_project_args) => {
+        BootstrapProject::process_bootstrap(bootstrap_project_args)
+      }
     }
   }
 
-  fn process_templates(config_provider: impl UserConfigProvider, process_templates: ProcessTemplatesArgs) -> ZatAction {
-    let user_config = config_provider.get_user_config(process_templates)?;
-    VerboseLogger::log_user_config(&user_config);
-
-    // Reads the .variables.zat-prompt file into TemplateVariables
-    let template_variable_provider = DefaultTemplateVariableProvider::new();
-    let template_variables = template_variable_provider.get_tokens(user_config.clone())?;
-    VerboseLogger::log_template_variables(&user_config, &template_variables);
-
-    // Ask for the user for the value of each variable
-    // Then verify all the variables supplied are correct
-    let template_config_validator = DefaultTemplateConfigValidator::new();
-    let template_variable_review = template_config_validator.validate(user_config.clone(), template_variables.clone());
-
-    match template_variable_review {
-      TemplateVariableReview::Accepted(vc) => {
-        VerboseLogger::log_user_supplied_variables(&user_config, &vc);
-        let user_variables = vc.user_variables;
-        let expand_filters = DefaultExpandFilters::new();
-        let tokenized_key_expanded_variables = expand_filters.expand_filers(template_variables, user_variables);
-
-        VerboseLogger::expanded_tokens(&user_config, &tokenized_key_expanded_variables);
-        DefaultProcessTemplates.process_templates(user_config.clone(), tokenized_key_expanded_variables)?;
-
-        // Run post-processor if one exists
-        ShellHook.run(&user_config)?;
-
-        Logger::coloured(
-          &s!("{}{}{}",
-            Logger::info_str("Extracted template to '"),
-            &user_config.target_dir.path.as_str(),
-            Logger::info_str("'")
-          ))
-      },
-      TemplateVariableReview::Rejected => Logger::warn("The user rejected the variables.")
-    }
-
-    Ok(())
-  }
-
-  fn process_boostrap(bootstrap_project_args: BootstrapProjectArgs) -> ZatAction {
-    // get the repository directory
-    let repository_directory = TargetDir::new(&bootstrap_project_args.repository_dir);
-    // verify that it doesn't exist
-    if repository_directory.does_exist() {
-      Err(ZatError::bootstrap_repository_dir_should_not_exist(&bootstrap_project_args.repository_dir))
-    } else {
-      // extract the sample files (return an error if this fails)
-      // display a message on how to run the template
-      todo!()
-    }
-  }
 }
